chore: refactor + lint markdown
