+++ .prettierrc.toml
@@ -1,0 +1,13 @@
+# .prettierrc.toml
+trailingComma = "es5"
+tabWidth = 8
+semi = false
+singleQuote = true
+
+[[overrides]]
+files = "*.md"
+[overrides.options]
+tabWidth = 2
+printWidth = 80
+proseWrap = "always"
+embeddedLanguageFormatting = "off"
