+++ keeper/keeper_test.go
@@ -25,8 +25,6 @@ import (
 
 	"cosmossdk.io/collections"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	capabilitytypes "github.com/cosmos/ibc-go/modules/capability/types"
-	host "github.com/cosmos/ibc-go/v8/modules/core/24-host"
 	"github.com/stretchr/testify/require"
 
 	"github.com/noble-assets/wormhole/types"
@@ -147,131 +145,4 @@ func TestParseAndVerifyVAA(t *testing.T) {
 	require.Equal(t, vaa2.Signatures, vaaResp.Signatures, "expected a different Signatures")
 	require.Equal(t, vaa2.Timestamp, vaaResp.Timestamp, "expected a different Timestamp")
 	require.Equal(t, vaa2.Version, vaaResp.Version, "expected a different Version")
-}
-
-func TestBindPort(t *testing.T) {
-	// ARRANGE
-	pk := mocks.PortKeeper{
-		Ports: make(map[string]bool),
-	}
-	sk := mocks.ScopedKeeper{
-		Capabilities: make(map[string]*capabilitytypes.Capability),
-	}
-	ics4w := mocks.ICS4Wrapper{}
-
-	ctx, k := mocks.NewWormholeKeeper(t, ics4w, pk, sk)
-
-	// ACT: Bind to port when no capability are stored.
-	err := k.BindPort(ctx)
-
-	// ASSERT
-	require.NoError(t, err, "expected no error when the capability associated with the port does not exist")
-
-	c, found := sk.Capabilities[host.PortPath(types.Port)]
-	require.True(t, found, "expected the capability to be in the state after binding")
-	require.Equal(t, &capabilitytypes.Capability{Index: uint64(3)}, c, "expected a different index for capability") // 3 is hardcoded in the mock
-
-	// ACT
-	err = k.BindPort(ctx)
-
-	// ASSERT
-	require.NoError(t, err, "expected no error when the capability is already registered")
-
-	// ARRANGE: Reset to initial conditions.
-	delete(sk.Capabilities, host.PortPath(types.Port))
-	// Setting the entry in the ports instructs the mock port to return a nil capability object.
-	pk.Ports[types.Port] = true
-
-	// ACT
-	err = k.BindPort(ctx)
-
-	// ASSERT
-	require.Error(t, err, "expected when claim capability returns an error")
-	require.ErrorContains(t, err, "could not claim port capability", "expected a different error")
-}
-
-func TestClaimCapability(t *testing.T) {
-	// ARRANGE
-	sk := mocks.ScopedKeeper{
-		Capabilities: make(map[string]*capabilitytypes.Capability),
-	}
-	pk := mocks.PortKeeper{}
-	ics4w := mocks.ICS4Wrapper{}
-
-	// ARRANGE
-	ctx, k := mocks.NewWormholeKeeper(t, ics4w, pk, sk)
-
-	// ACT
-	err := k.ClaimCapability(ctx, nil, "name")
-
-	// ASSERT
-	require.Error(t, err, "expected an error when the capability is nil")
-
-	// ARRANGE: Create a valid capability object.
-	capability := capabilitytypes.Capability{
-		Index: 3,
-	}
-
-	// ACT
-	err = k.ClaimCapability(ctx, &capability, "name")
-
-	c, found := sk.Capabilities["name"]
-	require.True(t, found, "expected the capability to be in the state")
-	require.Equal(t, &capability, c, "expected a different index for capability")
-
-	// ASSERT
-	require.NoError(t, err)
-}
-
-func TestPostMessage_Keeper(t *testing.T) {
-	// ARRANGE
-	pk := mocks.PortKeeper{
-		Ports: make(map[string]bool),
-	}
-	sk := mocks.ScopedKeeper{
-		Capabilities: make(map[string]*capabilitytypes.Capability),
-	}
-	ics4w := mocks.ICS4Wrapper{}
-
-	ctx, k := mocks.NewWormholeKeeper(t, ics4w, pk, sk)
-
-	// ACT
-	err := k.PostMessage(ctx, "", []byte{}, 0)
-
-	// ASSERT
-	require.Error(t, err, "expected an error when the wormchain channel is not set")
-	require.ErrorContains(t, err, "failed to get wormchain", "expected a different error")
-
-	// ARRANGE: Set the channel and cause an error with the packet data.
-	err = k.WormchainChannel.Set(ctx, "channel-0")
-	require.NoError(t, err, "expecting no error setting the wormhole channel")
-
-	// ACT
-	err = k.PostMessage(ctx, "", []byte{}, 0)
-
-	// ASSERT
-	require.Error(t, err, "expected an error because config is not set")
-	require.ErrorContains(t, err, "failed to get packet data")
-
-	// ARRANGE: Set the config but no capabilities. The GetCapability mock returns nil.
-	cfg := types.Config{ChainId: uint16(3)}
-	err = k.Config.Set(ctx, cfg)
-	require.NoError(t, err, "expected no error setting the config")
-
-	// ACT
-	signer := utils.TestAddress()
-	err = k.PostMessage(ctx, signer.Bech32, []byte{}, 0)
-
-	// ASSERT
-	require.Error(t, err, "expected an error when capability is nil and send packet is called")
-	require.ErrorContains(t, err, "failed to send packet")
-
-	// ARRANGE: Set a valid capability in the store.
-	sk.Capabilities[host.ChannelCapabilityPath(types.Port, "channel-0")] = &capabilitytypes.Capability{Index: uint64(3)}
-
-	// ACT
-	err = k.PostMessage(ctx, signer.Bech32, []byte("Hello from Noble"), 0)
-
-	// ASSERT
-	require.NoError(t, err)
 }
