+++ keeper/msg_server.go
@@ -52,7 +52,6 @@ func (k msgServer) SubmitVAA(ctx context.Context, msg *types.MsgSubmitVAA) (*typ
 	}
 
 	if vaa.EmitterChain != vaautils.ChainID(config.GovChain) || !bytes.Equal(vaa.EmitterAddress.Bytes(), config.GovAddress) {
-		// TODO: is this the error we want to return?
 		return nil, types.ErrNotGovernanceVAA
 	}
 	if vaa.GuardianSetIndex != config.GuardianSetIndex {
@@ -65,23 +64,17 @@ func (k msgServer) SubmitVAA(ctx context.Context, msg *types.MsgSubmitVAA) (*typ
 		return nil, errors.Wrap(err, "failed parsing the vaa payload")
 	}
 
-	if pkt.Chain != config.ChainId && pkt.Chain != 0 {
+	if pkt.Chain != 0 && pkt.Chain != config.ChainId {
 		return nil, errors.Wrap(types.ErrInvalidGovernanceVAA, "packet not meant for this chain")
 	}
 
 	switch pkt.Module {
 	case "Core":
 		err = k.HandleCoreGovernancePacket(ctx, pkt)
-		if err != nil {
-			err = errors.Wrap(err, "failed handling the core governance packet")
-		}
-		return &types.MsgSubmitVAAResponse{}, err
+		return &types.MsgSubmitVAAResponse{}, errors.Wrap(err, "failed handling the core governance packet")
 	case "IbcReceiver":
 		err = k.HandleIBCReceiverGovernancePacket(ctx, pkt)
-		if err != nil {
-			err = errors.Wrap(err, "failed handling the ibc receive governance packet")
-		}
-		return &types.MsgSubmitVAAResponse{}, err
+		return &types.MsgSubmitVAAResponse{}, errors.Wrap(err, "failed handling the ibc receive governance packet")
 	default:
 		return &types.MsgSubmitVAAResponse{}, errors.Wrapf(types.ErrUnsupportedGovernanceAction, "module: %s, type: %d", pkt.Module, pkt.Action)
 	}
